chore(argocd): add sync-wave annotations to monitoring applications
The sync-wave annotations have been added to the monitoring-kube-prometheus-stack-application.yaml and monitoring-kube-prometheus-stack-cdrs-application.yaml files. The sync-wave annotation value determines the order in which the applications are synced by Argo CD. The monitoring-kube-prometheus-stack-application.yaml has a sync-wave value of "1" and the monitoring-kube-prometheus-stack-cdrs-application.yaml has a sync-wave value of "0". This change improves the synchronization order of the monitoring applications.